Separate gdt bitness code
